Adds outfit history feature
Introduces a new component to display the user's outfit history,
including the date the outfit was used and the individual items.
The outfit history is stored in local storage.

Also, it fixes a bug where the "Sin gorra" option was not correctly
displayed when the hat carousel was at the first position.